Display S3 bucket name as well as path key to avoid confusion
I read `editorial-tools:explainer/24/artifacts.zip` to mean that the
artifact had been uploaded into an `editorial-tools` bucket - but it
did not! It was actually the first half of the first folder name in the
path - ie the first folder was called `editorial-tools:explainer`.

The new debug will show a full `s3://`-prefixed bucket name and path, to
be clearer.

This follows on from https://github.com/guardian/sbt-riffraff-artifact/pull/21#