fix: Add `aws-java-sdk-sts` as explicit dependency
This is to allow the use of `WebIdentityTokenCredentialsProvider` within GitHub Actions.

With https://github.com/guardian/actions-assume-aws-role we've unlocked the ability to use GHA for CI.

The blocker has previously been gaining AWS credentials.

https://github.com/guardian/actions-assume-aws-role allows GHA to assume a role to gain permissions to upload to Riff-Raff's buckets.

When trying to migrate a repository to GHA, we're greeted with this:

```
WebIdentityTokenCredentialsProvider: To use assume role profiles the aws-java-sdk-sts module must be on the class path.
```

Add `aws-java-sdk-sts` as explicit dependency to resolve.