fix: Improve branch lookup from GitHub Actions
The branch name within a GitHub Actions environment can be in one of two places:
  - `GITHUB_HEAD_REF` during PR builds
  - `GITHUB_REF` during a build

We read `GITHUB_HEAD_REF` and fallback to `GITHUB_REF`.

Either can be the empty string which is not very useful,
especially as Riff-Raff uses branch names for continuous and scheduled deployment.

This change guards against this possibility by checking an env var isn't empty and defaults to `"unknown-branch"`.